feat : 추가정보 입력 페이지 하나로 합치기
