+++ app/src/main/java/com/surveasy/surveasy/presentation/intro/RegisterEventType.kt
@@ -4,8 +4,7 @@ enum class RegisterEventType {
     TO_WARN,
     TO_TERM1,
     TO_TERM2,
-    TO_INPUT1,
-    TO_INPUT2,
+    TO_INPUT,
     TO_BACK,
     TO_MAIN
 }
