+++ app/src/main/java/com/surveasy/surveasy/presentation/intro/RegisterInputFragment.kt
@@ -7,14 +7,14 @@ import android.widget.ArrayAdapter
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
 import com.surveasy.surveasy.R
-import com.surveasy.surveasy.databinding.FragmentRegisterInput2Binding
+import com.surveasy.surveasy.databinding.FragmentRegisterInputBinding
 import com.surveasy.surveasy.presentation.base.BaseFragment
 import com.surveasy.surveasy.presentation.main.MainActivity
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class RegisterInput2Fragment :
-    BaseFragment<FragmentRegisterInput2Binding>(R.layout.fragment_register_input2) {
+class RegisterInputFragment :
+    BaseFragment<FragmentRegisterInputBinding>(R.layout.fragment_register_input) {
     private val viewModel: RegisterViewModel by viewModels()
 
     override fun initData() = Unit
@@ -34,8 +34,8 @@ class RegisterInput2Fragment :
 
     override fun initView() = with(binding) {
         initBankSpinner()
+        initInflowPathSpinner()
         vm = viewModel
-        lifecycleOwner = viewLifecycleOwner
     }
 
     private fun initBankSpinner() = with(binding) {
@@ -56,6 +56,32 @@ class RegisterInput2Fragment :
                     id: Long
                 ) {
                     viewModel.setBank(bankList[position])
+                }
+
+                override fun onNothingSelected(p0: AdapterView<*>?) {
+                }
+            }
+        }
+    }
+
+    private fun initInflowPathSpinner() = with(binding) {
+        val inflowPathList = resources.getStringArray(R.array.inflowPath)
+        val inflowPathAdapter = ArrayAdapter(
+            requireContext(),
+            androidx.appcompat.R.layout.support_simple_spinner_dropdown_item,
+            inflowPathList
+        )
+
+        sInflow.apply {
+            adapter = inflowPathAdapter
+            onItemSelectedListener = object : AdapterView.OnItemSelectedListener {
+                override fun onItemSelected(
+                    parent: AdapterView<*>?,
+                    view: View?,
+                    position: Int,
+                    id: Long
+                ) {
+                    viewModel.setInflow(inflowPathList[position])
                 }
 
                 override fun onNothingSelected(p0: AdapterView<*>?) {
