+++ app/src/main/res/layout/fragment_register_agree.xml
@@ -209,7 +209,7 @@
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:indeterminate="false"
-            android:max="5"
+            android:max="4"
             android:minHeight="50dp"
             android:progress="1"
             android:theme="@style/progressBar"
