+++ app/src/main/res/layout/fragment_register_warn.xml
@@ -227,7 +227,7 @@
             android:layout_height="wrap_content"
             android:layout_weight="1"
             android:indeterminate="false"
-            android:max="5"
+            android:max="4"
             android:minHeight="50dp"
             android:progress="2"
             android:theme="@style/progressBar"
@@ -243,7 +243,7 @@
             android:layout_marginHorizontal="5sp"
             android:background="@drawable/register_button"
             android:fontFamily="@font/main_font_b"
-            android:onClick="@{() -> vm.navigateRegisterPages(RegisterEventType.TO_INPUT1)}"
+            android:onClick="@{() -> vm.navigateRegisterPages(RegisterEventType.TO_INPUT)}"
             android:text="@string/agree2_btn"
             android:textColor="@color/white"
             android:textSize="15sp"
